+++ Quaver.API/Replays/ReplayFrame.cs
@@ -5,7 +5,7 @@
         /// <summary>
         ///     The time in the replay since the last frame.
         /// </summary>
-        public float Time { get; }
+        public int Time { get; }
 
         /// <summary>
         ///     The keys that were pressed during this frame.
@@ -17,7 +17,7 @@
         /// </summary>
         /// <param name="time"></param>
         /// <param name="keys"></param>
-        public ReplayFrame(float time, ReplayKeyPressState keys)
+        public ReplayFrame(int time, ReplayKeyPressState keys)
         {
             Time = time;
             Keys = keys;
